feat(dedicated-servers): column visibility in views and switch views
ref: #MANAGER-20621
