+++ packages/manager/apps/dedicated-servers/src/components/manageView/types.ts
@@ -1,5 +1,8 @@
+import { VisibilityState } from '@tanstack/react-table';
+
 export type ViewType = {
   name: string;
   id: string;
   default?: boolean;
+  columnVisibility?: VisibilityState;
 };
